Move GetTempPath() to testutil.
Conflicts:
	src/Makefile.test.include
	src/test/alert_tests.cpp
	src/test/test_bitcoin.cpp
	src/test/testutil.cpp
	src/test/testutil.h
	src/util.cpp
