+++ src/Makefile.test.include
@@ -85,6 +85,7 @@ BITCOIN_TESTS =\
   test/skiplist_tests.cpp \
   test/test_bitcoin.cpp \
   test/test_bitcoin.h \
+  test/testutil.cpp \
   test/testutil.h \
   test/thinblock_tests.cpp \
   test/thinblockbuilder_tests.cpp \
